norm対応とstatus code処理の修正
