+++ src/utils.c
@@ -3,15 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: koseki.yusuke <koseki.yusuke@student.42    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/16 12:07:35 by koseki.yusu       #+#    #+#             */
-/*   Updated: 2024/08/29 01:02:43 by sakitaha         ###   ########.fr       */
+/*   Updated: 2024/09/07 23:07:36 by koseki.yusu      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "utils.h"
+#include "xlibc.h"
 
 void	ft_bzero(void *s, size_t n)
 {
@@ -136,4 +137,33 @@ char	*ft_strncpy(char *dest, const char *src, size_t n)
 		i++;
 	}
 	return (dest);
+}
+
+char	*ft_strdup(char *src)
+{
+	char	*p;
+	long	len;
+
+	if (src == NULL)
+		return (NULL);
+	len = ft_strlen(src);
+	p = (char *)xmalloc(sizeof(char) * (len + 1));
+	if (!p)
+		return (0);
+	ft_strcpy(p, src);
+	return (p);
+}
+
+char	*ft_strndup(char *src, long len)
+{
+	char	*p;
+
+	if (src == NULL)
+		return (NULL);
+	p = (char *)xmalloc(sizeof(char) * (len + 1));
+	if (!p)
+		return (NULL);
+	ft_strncpy(p, src, len);
+	p[len] = '\0';
+	return (p);
 }
